requirements.txt
langgraph
langchain
flask
flask-cors
torch
PyPDF2
python-docx
transformers
huggingface_hub
python-dotenv